GIF export speedup
This PR was originally made by @novhack for Pixelorama (https://github.com/Orama-Interactive/Pixelorama/pull/696), but the changes weren't brought in this repo.

Note that I had to not keep the changes made in the `add_image_data_block()` method in exported.gd that the original PR had, because the resulting gif file seemed invalid. I'm not sure what the problem is, but even without this change the speedup is still there.

Using the default images of the project:
Gif exporting time before this PR: 1.70701233333333
Gif exporting time after this PR: 1.01357166666667